requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.9.0
pydantic-settings==2.5.0
requests==2.31.0
python-dotenv==1.0.0
jinja2>=3.1.2
alembic==1.12.1
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
